backend/app/api/routes/recolector_logs/requirements.txt
fastapi
pydantic
asyncpg
python-dotenv
uvicorn
httpx

